# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/index": {
            "get": {
                "description": "这是一个测试方法",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger测试"
                ],
                "summary": "测试",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用于用户登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "无需登陆"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "登陆参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"用户登陆成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"用户登陆失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_create": {
            "post": {
                "description": "用于用户创建商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "创建商品",
                "parameters": [
                    {
                        "description": "创建商品需要的参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProductCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品创建失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_delete": {
            "delete": {
                "description": "用于用户删除商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "删除商品",
                "parameters": [
                    {
                        "description": "需要删除商品的id",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProductDeleteParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_detail": {
            "get": {
                "description": "用户可以根据输入商品id查看相应商品详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看商品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "productId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品详情查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品详情查询失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_flow": {
            "get": {
                "description": "接受商品id列表，返回对应商品信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看商品推荐流",
                "parameters": [
                    {
                        "type": "string",
                        "name": "key_word",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品流查看成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品流查看失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_list": {
            "get": {
                "description": "用户可以根据输入商品category_id查看相应商品列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "categoryId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品列表查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品列表查询失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_modify": {
            "put": {
                "description": "用于用户修改商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "修改商品",
                "parameters": [
                    {
                        "description": "需要修改商品的参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProductModifyParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品修改失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_search": {
            "get": {
                "description": "根据用户输入关键字搜索商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "搜索商品",
                "parameters": [
                    {
                        "type": "string",
                        "name": "key_word",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品搜索成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品搜索失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shop_create": {
            "post": {
                "description": "卖家输入店铺信息，创建店铺",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "创建店铺",
                "parameters": [
                    {
                        "description": "创建商铺参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShopCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"店铺创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"店铺创建失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shop_delete": {
            "delete": {
                "description": "用于卖家删除店铺",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "删除店铺",
                "parameters": [
                    {
                        "description": "删除店铺参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShopDeleteParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"店铺删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"店铺删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shop_list": {
            "get": {
                "description": "卖家查看自己的商铺列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看店铺列表",
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"店铺列表查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"店铺列表查询失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shop_modify": {
            "put": {
                "description": "用于卖家修改店铺信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "修改店铺信息",
                "parameters": [
                    {
                        "description": "修改店铺参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShopModifyParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"店铺信息修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"店铺新秀修改失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shop_product": {
            "get": {
                "description": "如果是买家则展示所有商品，卖家则可以根据商品状态查询对应商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看店铺下商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "需要查看的店铺id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "需要查看的商品状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"商品列表查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"商品列表查询失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user_info": {
            "get": {
                "description": "用于已登陆的用户查看个人信息，包括用户名、邮箱、头像",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "需要登陆"
                ],
                "summary": "查看用户信息",
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"用户信息查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"用户信息查询失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ProductCreateParam": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "shop_id": {
                    "type": "integer"
                },
                "store": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.ProductDeleteParam": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "model.ProductModifyParam": {
            "type": "object",
            "required": [
                "id",
                "name",
                "price"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "shop_id": {
                    "type": "integer"
                },
                "store": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.ShopCreateParam": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "detail": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.ShopDeleteParam": {
            "type": "object",
            "properties": {
                "shop_id": {
                    "type": "integer"
                }
            }
        },
        "model.ShopModifyParam": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.UserLoginParam": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}